pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "engineered-genetic-circuits-in-3d-space"
version = "0.0.1"
requires-python = ">=3.7"
dependencies = [
	     "biopython",
	     "numpy",
	     "pandas",
	     "matplotlib",
	     "seaborn"
]
license = {text = "MIT License"}
